Stop prepareTestLoader from clobbering prepareTestResult.
Returning a new TestLoader from prepareTestLoader prevented nose from
calling prepareTestResult, so docstrings would still appear. (I'm not
sure exactly why this occurs - I haven't found the cause in the nose
code, although there are a few parts that look to contribute to it.)
Changed prepareTestLoader to patch the passed-in loader instead.